Develop

+++ resources/views/categories/index.blade.php
@@ -5,7 +5,7 @@
         'title' => 'Категории',
         'route' => 'categories',
         'breadcrumbs' => 'Категории',
-        'add_route' => 'categories.create',
+        'add_route' => $canCreateProduct ? 'categories.create' : null
     ])
 
     <div class="row">
@@ -25,12 +25,16 @@
                             <td>{{ $category->name }}</td>
                             <td>
                                 <a href="{{ route('categories.show', $category) }}" class="btn btn-primary">Просмотр</a>
-                                <a href="{{ route('categories.edit', $category) }}" class="btn btn-warning">Редактировать</a>
-                                <x-form action="{{ route('categories.delete', $category) }}" method="DELETE"
-                                    style="display: inline-block;">
-                                    <button onclick="return confirm('Вы уверены?')" type="submit"
-                                        class="btn btn-danger">Удалить</button>
-                                </x-form>
+                                @can('update', \App\Models\Product::class)
+                                    <a href="{{ route('categories.edit', $category) }}" class="btn btn-warning">Редактировать</a>
+                                @endcan
+                                @can('delete', \App\Models\Product::class)
+                                    <x-form action="{{ route('categories.delete', $category) }}" method="DELETE"
+                                        style="display: inline-block;">
+                                        <button onclick="return confirm('Вы уверены?')" type="submit"
+                                            class="btn btn-danger">Удалить</button>
+                                    </x-form>
+                                @endcan
                             </td>
                         </tr>
                     @endforeach
