+++ resources/views/divisions/index.blade.php
@@ -5,7 +5,7 @@
         'title' => 'Подразделения',
         'route' => 'divisions',
         'breadcrumbs' => 'Подразделения',
-        'add_route' => 'divisions.create',
+        'add_route' => $canCreateProduct ? 'divisions.create' : null,
     ])
 
     <div class="row">
@@ -25,12 +25,16 @@
                             <td>{{ $division->name }}</td>
                             <td>
                                 <a href="{{ route('divisions.show', $division) }}" class="btn btn-primary">Просмотр</a>
+                                @can('update', \App\Models\Product::class)
                                 <a href="{{ route('divisions.edit', $division) }}" class="btn btn-warning">Редактировать</a>
-                                <x-form action="{{ route('divisions.delete', $division) }}" method="DELETE"
-                                    style="display: inline-block;">
-                                    <button onclick="return confirm('Вы уверены?')" type="submit"
-                                        class="btn btn-danger">Удалить</button>
-                                </x-form>
+                                @endcan
+                                @can('delete', \App\Models\Product::class)
+                                    <x-form action="{{ route('divisions.delete', $division) }}" method="DELETE"
+                                        style="display: inline-block;">
+                                        <button onclick="return confirm('Вы уверены?')" type="submit"
+                                            class="btn btn-danger">Удалить</button>
+                                    </x-form>
+                                @endcan
                             </td>
                         </tr>
                     @endforeach
